requirements.txt
a2a-sdk==0.3.10
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
fastapi==0.120.3
genai-prices==0.0.35
google-api-core==2.28.1
google-auth==2.42.1
google-genai==1.47.0
googleapis-common-protos==1.71.0
griffe==1.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.0
logfire-api==4.14.2
opentelemetry-api==1.38.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic-ai-slim==1.9.1
pydantic-graph==1.9.1
pydantic_core==2.41.4
python-dotenv==1.2.1
redis==7.0.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
sse-starlette==3.0.3
starlette==0.49.1
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
zipp==3.23.0
